+++ astraea/src/lib.rs
@@ -1,0 +1,3 @@
+#![feature(array_chunks)]
+pub mod compiler;
+pub mod tree;
