+++ nonlocality_host/Cargo.toml
@@ -6,13 +6,5 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-dogbox_blob_layer = { path ="../dogbox/dogbox_blob_layer" }
+astraea = { path ="../astraea" }
 tokio = {version = "1", features = ["rt-multi-thread", "macros", "time", "sync"]}
-tokio-stream = "0"
-async-stream = "0"
-futures = "0"
-futures-util = "0"
-futures-core = "0"
-sha3 = "0"
-serde = "1"
-postcard = {version = "1", features = ["alloc"]}
