+++ nonlocality_host/src/main.rs
@@ -1,19 +1,18 @@
 #![feature(array_chunks)]
-mod lib;
-use lib::make_beginning_of_time;
-use lib::make_delay;
-use lib::make_seconds;
-use lib::make_text_in_console;
-use lib::reduce_expression_without_storing_the_final_result;
-use lib::ActualConsole;
-use lib::DelayService;
-use lib::Identity;
-use lib::InMemoryValueStorage;
-use lib::ReduceExpression;
-use lib::ServiceRegistry;
-use lib::StoreValue;
-use lib::TypeId;
-use lib::Value;
+use astraea::tree::make_beginning_of_time;
+use astraea::tree::make_delay;
+use astraea::tree::make_seconds;
+use astraea::tree::make_text_in_console;
+use astraea::tree::reduce_expression_without_storing_the_final_result;
+use astraea::tree::ActualConsole;
+use astraea::tree::DelayService;
+use astraea::tree::Identity;
+use astraea::tree::InMemoryValueStorage;
+use astraea::tree::ReduceExpression;
+use astraea::tree::ServiceRegistry;
+use astraea::tree::StoreValue;
+use astraea::tree::TypeId;
+use astraea::tree::Value;
 use std::collections::BTreeMap;
 use std::sync::Arc;
 use std::sync::Mutex;
